Remove of redundant interfaces Added logic in Transmissin class Refactoring of multiple classes

+++ RpmMeter.Contracts/Entities/Piston.cs
@@ -8,7 +8,7 @@ namespace RpmMeter.Contracts.Entities
 {
     public class Piston
     {
-        private decimal Position { get; set; }
+        private int Position { get; set; }
 
     }
 }
